+++ sirius-ui/src/server/auth.ts
@@ -64,8 +64,7 @@ const getNextAuthUrl = (req?: any): string => {
  * @see https://next-auth.js.org/configuration/options
  */
 export const authOptions: NextAuthOptions = {
-  secret:
-    process.env.NEXTAUTH_SECRET || "change-this-secret-in-production-please",
+  secret: env.NEXTAUTH_SECRET || undefined,
   session: {
     strategy: "jwt",
     maxAge: 100 * 365 * 24 * 60 * 60, // 100 years in seconds - effectively indefinite